Shortlisting valid indicators based on Sharpe Ratio
Using a separate backtest period from Oct 2019-Dec 2019 to determine the validity of individual technical indicators on each security. These technical indicators are deemed to be valid if they produce positive sharpe ratio over the testing period.  After which, the valid indicators for each security are used in majority voting for the actual backtest period